+++ client/src/ApplicationViews.js
@@ -1,16 +1,27 @@
 import { Navigate, Route, Routes } from "react-router-dom";
 import { Login } from "./components/auth/Login";
-import { UserVideos } from "./components/UserVideos";
+import { UserVideos } from "./components/user/UserVideos";
 import { Register } from "./components/auth/Register";
+import { SearchBar } from "./components/search/SearchBar";
 
-export const ApplicationViews = ({ isLoggedIn, role }) => {
+export const ApplicationViews = ({ isLoggedIn, role, currentUser }) => {
 	return (
 		<main>
 			<Routes>
 				<Route path="/">
 					<Route
 						index
-						element={isLoggedIn ? <UserVideos /> : <Navigate to="/login" />}
+						element={
+							isLoggedIn ? (
+								<UserVideos currentUser={currentUser} />
+							) : (
+								<Navigate to="/login" />
+							)
+						}
+					/>
+					<Route
+						path="searchbar"
+						element={isLoggedIn ? <SearchBar /> : <Navigate to="/login" />}
 					/>
 					<Route path="login" element={<Login />} />
 					<Route path="register" element={<Register />} />
